requirements.txt
django==1.11
djangorestframework
coreapi
markdown
lxml
beautifulsoup4
django-tagging
django-filter
requests
